# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Habarchy API
  version: 0.5.0
  description: |
    Habarchy (Habarçy) is a self-hosted, multi-tenant notification gateway: one API to send
    SMS, email, push (Firebase FCM) and Telegram messages with queues, retries, provider
    fallback, templates, delivery status, webhooks and an admin panel.

    Two APIs share this document:
    * **Public API** `/api/v1` — called by client applications with an API key (`X-Api-Key`),
      optionally HMAC-signed (`X-Signature`, `X-Timestamp`). Keys carry scopes.
    * **Admin API** `/api/admin` — used by the web and mobile admin apps with a JWT
      (`Authorization: Bearer`). Project routes are role based (owner > admin > developer > viewer).

    Every response is wrapped in an envelope `{ "data": ..., "meta": ..., "error": {"code","message","details"} }`.
servers:
  - url: http://localhost:8080
...
security: []
paths:
...
  /api/v1/me:
    get:
      tags: [public-usage]
      summary: Identify the API key and its project
      security: [{ ApiKey: [] }]
      responses:
        "200":
          description: Key info
          content: { application/json: { schema: { $ref: "#/components/schemas/Envelope" } } }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/gateway/v1/me:
    get:
      tags: [gateway]
      summary: Who am I (phone pairing check)
      security: [{ GatewayKey: [] }]
      responses:
        "200": { description: Provider, project and pending count }
        "401": { description: Invalid gateway key, disabled provider or suspended project }
  /api/gateway/v1/outbox:
    get:
      tags: [gateway]
      summary: Lease pending SMS (long poll)
      description: Returns up to `limit` SMS the phone must send and marks them leased. Waits up to `wait` seconds (max 25) when the outbox is empty.
      security: [{ GatewayKey: [] }]
      parameters:
        - { name: wait, in: query, schema: { type: integer, default: 20, minimum: 0, maximum: 25 } }
        - { name: limit, in: query, schema: { type: integer, default: 5, minimum: 1, maximum: 20 } }
      responses:
        "200":
          description: Items to send
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { type: array, items: { $ref: "#/components/schemas/OutboxItem" } } }
  /api/gateway/v1/outbox/{outbox_id}/result:
    parameters: [{ name: outbox_id, in: path, required: true, schema: { type: string, format: uuid } }]
    post:
      tags: [gateway]
      summary: Report sent / failed / delivered for an outbox item
      security: [{ GatewayKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [sent, failed, delivered] }
                error_code: { type: string, description: "Android SmsManager result: generic_failure, radio_off, null_pdu, no_service, limit_exceeded, …" }
                error_message: { type: string }
                parts: { type: integer }
      responses:
        "200": { description: "`{id, status, applied}`" }
        "404": { description: Unknown outbox id for this phone }
  /api/gateway/v1/heartbeat:
    post:
      tags: [gateway]
      summary: Phone heartbeat (battery, network, operator, app version)
      security: [{ GatewayKey: [] }]
      requestBody: { content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses:
        "200": { description: "`{ok, pending, server_time}`" }
  /api/gateway/v1/inbound:
    post:
      tags: [gateway]
      summary: Forward an SMS the phone received (fires the sms.inbound webhook)
      security: [{ GatewayKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from, text]
              properties: { from: { type: string }, text: { type: string }, received_at: { type: string, format: date-time } }
      responses:
        "201": { description: Stored }
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: "`hb_live_...` or `hb_test_...`. Optionally sign requests with `X-Timestamp` and `X-Signature` (see docs/auth.md)."
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    GatewayKey:
      type: apiKey
      in: header
      name: X-Gateway-Key
      description: "Pairing key of an `android_sms` provider (Habarçy Gateway app)."
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
...
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta: { type: object, additionalProperties: true }
        error: { $ref: "#/components/schemas/APIError" }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/APIError" }
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [validation_failed, unauthorized, forbidden_scope, not_found, quota_exceeded, rate_limited, duplicate_request, provider_unavailable, invalid_recipient, conflict, internal_error]
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    OutboxItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        message_id: { type: string, format: uuid }
        to: { type: string }
        text: { type: string }
        sim_slot: { type: integer, description: "-1 default SIM, 0 or 1" }
        expires_at: { type: string, format: date-time }
...